requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pytest==7.4.4
watchdog==3.0.0